fix::> removed the tempfile and moved all the changes to types_transaction_receipt.go
